requirements.txt
eventlet==0.25.1
python-socketio==4.5.1
